fixing unittests
